requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-socketio[asgi]>=5.11.1
starlette>=0.27.0